requirements.txt
pdfplumber
PyPDF
mysql-connector-python
spacy
traits